requirements.txt
Django
djangorestframework
google-generativeai
python-dotenv
requests